revert attribute check to ir_generator
