+++ pkg/sql/ir_generator_test.go
@@ -58,7 +58,9 @@ func TestGenerateTrainStmt(t *testing.T) {
 	r, e := parser.ParseStatement("mysql", normal)
 	a.NoError(e)
 
-	trainStmt, err := generateTrainStmt(r.SQLFlowSelectStmt)
+	trainStmt, err := generateTrainStmt(r.SQLFlowSelectStmt, true)
+	a.Error(err)
+	trainStmt, err = generateTrainStmt(r.SQLFlowSelectStmt, false)
 	a.NoError(err)
 	a.Equal("DNNClassifier", trainStmt.Estimator)
 	a.Equal(`SELECT c1, c2, c3, c4 FROM my_table
@@ -204,7 +206,7 @@ func TestGenerateTrainStmtWithTypeCheck(t *testing.T) {
 	r, e := parser.ParseStatement("mysql", normal)
 	a.NoError(e)
 
-	trainStmt, err := generateTrainStmt(r.SQLFlowSelectStmt)
+	trainStmt, err := generateTrainStmt(r.SQLFlowSelectStmt, true)
 	a.NoError(err)
 	a.Equal("DNNClassifier", trainStmt.Estimator)
 	a.Equal("SELECT c1, c2, c3, c4 FROM my_table\n	", trainStmt.Select)
@@ -262,7 +264,7 @@ func TestGenerateTrainStmtModelZoo(t *testing.T) {
 	r, e := parser.ParseStatement("mysql", normal)
 	a.NoError(e)
 
-	trainStmt, err := generateTrainStmt(r.SQLFlowSelectStmt)
+	trainStmt, err := generateTrainStmt(r.SQLFlowSelectStmt, false)
 	a.NoError(err)
 	a.Equal("a_data_scientist/regressors:v0.2", trainStmt.ModelImage)
 	a.Equal("MyDNNRegressor", trainStmt.Estimator)
@@ -421,7 +423,7 @@ func bucketColumnParserTestMain(bucketStr string) error {
 		return err
 	}
 
-	trainStmt, err := generateTrainStmt(pr[0].SQLFlowSelectStmt)
+	trainStmt, err := generateTrainStmt(pr[0].SQLFlowSelectStmt, false)
 	if err != nil {
 		return err
 	}
